requirements.txt
Flask
Bootstrap-Flask
pyserial
